- UPdated admin controller logic

+++ src/Controllers/AdminController.cs
@@ -263,6 +263,7 @@ namespace AzureNamingTool.Controllers
         [Route("[action]/{id}")]
         public async Task<IActionResult> GetGeneratedName(int id)
         {
+            ServiceResponse serviceResponse = new();
             try
             {
                 serviceResponse = await GeneratedNamesService.GetItem(id);
@@ -286,22 +287,51 @@ namespace AzureNamingTool.Controllers
         /// <summary>
         /// This function will delete the generated names data by ID.
         /// </summary>
+        /// <param name="adminpassword">string - Admin password</param>
         /// <param name="id">int - Generated Name id</param>
         /// <returns>bool - PASS/FAIL</returns>
         [HttpDelete]
         [Route("[action]/{id}")]
-        public async Task<IActionResult> DeleteGeneratedName(int id)
+        public async Task<IActionResult> DeleteGeneratedName([BindRequired][FromHeader(Name = "AdminPassword")] string adminpassword, int id)
         {
+            ServiceResponse serviceResponse = new();
             try
             {
-                serviceResponse = await GeneratedNamesService.DeleteItem(id);
-                if (serviceResponse.Success)
+                if (GeneralHelper.IsNotNull(adminpassword))
                 {
-                    return Ok(serviceResponse.ResponseObject);
+                    if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
+                    {
+                        // Get the item details
+                        serviceResponse = await GeneratedNamesService.GetItem(id);
+                        if (serviceResponse.Success)
+                        {
+                            GeneratedName item = (GeneratedName)serviceResponse.ResponseObject!;
+                            serviceResponse = await GeneratedNamesService.DeleteItem(id);
+                            if (serviceResponse.Success)
+                            {
+                                AdminLogService.PostItem(new AdminLogMessage() { Source = "API", Title = "INFORMATION", Message = "Generated Name (" + item.ResourceName + ") deleted." });
+                                CacheHelper.InvalidateCacheObject("GeneratedName");
+                                return Ok("Generated Name (" + item.ResourceName + ") deleted.");
+                            }
+                            else
+                            {
+                                return BadRequest(serviceResponse.ResponseObject);
+                            }
+                        }
+                        else
+                        {
+                            return BadRequest(serviceResponse.ResponseObject);
+                        }
+                    }
+                    else
+                    {
+                        return Ok("FAILURE - Incorrect Global Admin Password.");
+                    }
+
                 }
                 else
                 {
-                    return BadRequest(serviceResponse.ResponseObject);
+                    return Ok("FAILURE - You must provide the Global Admin Password.");
                 }
             }
             catch (Exception ex)
